Relax setuptools version constraint.
Requiring setuptools<31 now breaks travis builds, so relax the constraint.
This was bound to happen sooner or later.
Allowing the latest setuptools version has only one drawback:
it obliges people using wheels built with recent versions to
always use odoo-autodiscover>=2 on Odoo<11.
This is the safest and easiest anyway and it is best practice
so most people should be unaffected.