[DOC] Require suffix for XML file names
Currently, there is a discrepancy between the documentation and the
module template, regarding how XML files should be named. The
documentation says they should be without suffix
(views/res_partner.xml), but the model template says it should be with
suffix (views/res_partner_view.xml).

To be consistent, this modifies the documentation so it matches the
module template.

Closes #388